simple design changes made and modified delete functionality

+++ admin/pharmacist/load.php
@@ -20,11 +20,12 @@ foreach($data as $value)
                 <td><?php echo $value['phone'] ?></td>
                 <td><?php echo $value['address'] ?></td>
                 <td><a href="update.php?editid=<?php echo $value['id']; ?>" class="edit btn-update btn-big ">edit</a></td>
-                <td><a href="delete.php?deleteid=<?php echo $value['id']; ?>" class="delete btn-delete btn-big">delete</a></td>
+                <td><a onclick="return confirm('Are you sure you want to delete this Pharmacist?');"  href="delete.php?deleteid=<?php echo $value['id']; ?>" class="delete btn-delete btn-big">delete</a></td>
               </tr>
               <?php 
                 }
 }
-
-
-?>
+else{
+  ?>
+  <td colspan="8" class="no-record">No records found</td>
+<?php } ?>
